check for pkgconfig before installing and install smpeg separately

+++ installer.sh
@@ -2,18 +2,17 @@
 
 unameOut="$(uname -s)"
 case "${unameOut}" in
-    Linux*)     MACHINE=Linux;;
-    Darwin*)    MACHINE=Mac;;
-    CYGWIN*)    MACHINE=Cygwin;;
-    MINGW*)     MACHINE=MinGw;;
+    Linux*)     MACHINE=Linux ;;
+    Darwin*)    MACHINE=Mac ;;
+    CYGWIN*)    MACHINE=Cygwin ;;
+    MINGW*)     MACHINE=MinGw ;;
     *)          MACHINE="UNKNOWN:${unameOut}"
 esac
 echo Machine identified as $MACHINE
 
 if [ $MACHINE = Linux ]; then
     USER=`whoami`
-    if [ $USER = "root" ]
-        then
+    if [ $USER = "root" ]; then
         echo "Running as ROOT"
     else
         echo "***********YOU ARE NOT ROOT************"
@@ -25,18 +24,20 @@ DIR=`pwd`
 
 function finish {
     echo "Cleaning up..."
-    rm $DIR/pkgconfig.tgz || true
-    rm $DIR/SDL2-2.0.7.tar.gz || true
-    rm $DIR/freetype-2.8.tar.gz || true
-    rm $DIR/SDL2_ttf-2.0.14.tar.gz || true
-    rm $DIR/SDL2_image-2.0.2.tar.gz || true
-    rm $DIR/SDL2_mixer-2.0.2.tar.gz || true
+    rm -f $DIR/pkgconfig.tgz || true
+    rm -f $DIR/SDL2-2.0.7.tar.gz || true
+    rm -f $DIR/freetype-2.8.tar.gz || true
+    rm -f $DIR/SDL2_ttf-2.0.14.tar.gz || true
+    rm -f $DIR/SDL2_image-2.0.2.tar.gz || true
+    rm -f $DIR/SDL2_mixer-2.0.2.tar.gz || true
+    rm -f $DIR/smpeg2-2.0.0.tar.gz || true
     rm -rf $DIR/pkg-config-0.28 || true
     rm -rf $DIR/SDL2_image-2.0.2  || true
     rm -rf $DIR/freetype-2.8 || true
     rm -rf $DIR/SDL2_ttf-2.0.14  || true
     rm -rf $DIR/SDL2-2.0.7 || true
     rm -rf $DIR/SDL2_mixer-2.0.2 || true
+    rm -rf $DIR/smpeg2-2.0.0 || true
     rm -rf $DIR/glog || true
     rm -rf $DIR/yaml-cpp || true
 }
@@ -44,7 +45,7 @@ function finish {
 trap finish EXIT
 
 #pkg-config
-if [ $MACHINE = Mac ]; then
+if [[ $MACHINE = Mac && ! -f /usr/local/bin/x86_64-apple-darwin17.3.0-pkg-config ]]; then
     curl https://pkg-config.freedesktop.org/releases/pkg-config-0.28.tar.gz -o pkgconfig.tgz
     tar -zxf pkgconfig.tgz && cd pkg-config-0.28
     ./configure --with-internal-glib && make install
@@ -89,13 +90,15 @@ cd $DIR
 wget https://www.libsdl.org/projects/SDL_mixer/release/SDL2_mixer-2.0.2.tar.gz
 tar xzf SDL2_mixer-2.0.2.tar.gz
 cd SDL2_mixer-2.0.2
-
-#SMPEG
-cd external/smpeg2-2.0.0
 ./configure
 make -j4
 make install -j4
-cd ../../
+cd $DIR
+
+#SMPEG
+wget https://www.libsdl.org/projects/smpeg/release/smpeg2-2.0.0.tar.gz
+tar xzf smpeg2-2.0.0.tar.gz
+cd smpeg2-2.0.0
 ./configure
 make -j4
 make install -j4
@@ -104,8 +107,8 @@ cd $DIR
 # glog
 git clone https://github.com/google/glog.git
 cd glog
-mkdir build
-cd build
+mkdir tmp
+cd tmp
 cmake ..
 make -j4
 make install -j4
